feat(choose,filter): --input-delimiter --output-delimiter
allows to change how content from stdin is used, and how results are
printed.

one could get around it piping into and from `tr`, but results aren't
quite right, especially when `tr '\n' ','` for example, as it'll add an
extra `,` in the end of the string.

This makes things a bit cleaner, hopefully.

closes #274