fix(file): use height parameter as max height
While adding the padding parameter to file, the component has been
refactored to use `m.filepicker.SetHeight` instead of the deprecated
property `m.filepicker.Height`. Since this change, the size has
been calculated based of the window height, overriding the value
passed with `--height`.
This adds a new `maxHeight` property to constrain the maximum
height of the file picker to the value passed by `--height`.

Refs: 6045525, #969